+++ lib/reddit.dart
@@ -14,6 +14,10 @@ import "package:oauth2/src/handle_access_token_response.dart";
 import "src/fullname.dart";
 
 
+export "src/fullname.dart";
+export "src/reddit_uri.dart";
+
+
 part "src/reddit.dart";
 part "src/subreddit.dart";
 part "src/listing.dart";